master/pyproject.toml
[build-system]
requires = [
    "setuptools>=80",
    "setuptools-scm>=9",
]
build-backend = "setuptools.build_meta"

[project]
name = "buildbot"
requires-python = ">=3.9"
authors = [
    { name = "Brian Warner", email = "warner-buildbot@lothar.com" },
]
maintainers = [
    { name = "Dustin J. Mitchell", email = "dustin@v.igoro.us" },
]
description = "The Continuous Integration Framework"
readme = "README.rst"
license = {text = "GPL-2.0-or-later"}
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: No Input/Output (Daemon)",
    "Environment :: Web Environment",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Build Tools",
    "Topic :: Software Development :: Testing",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]
dependencies = [
    'setuptools >= 8.0',
    'Twisted >= 24.7.0',
    'treq >= 20.9',
    'Jinja2 >= 2.1',
    'msgpack >= 0.6.0',
    "croniter >= 1.3.0",
    # required for tests, but Twisted requires this anyway
    'zope.interface >= 4.1.1',
    'sqlalchemy >= 1.4.0',
    'alembic >= 1.6.0',
    'python-dateutil>=1.5',
    "txaio >= 2.2.2",
    "autobahn >= 0.16.0",
    'packaging',
    'PyJWT',
    'pyyaml',
    'unidiff >= 0.7.5',
    # buildbot_windows_service needs pywin32
    'pywin32; platform_system=="Windows"',
]

dynamic = [
    "version",
    "scripts",
    "entry-points",
]
[project.optional-dependencies]
test = [
    "ruff",
    # http client libraries
    'treq',
    'txrequests',
    # pypugjs required for custom templates tests
    'pypugjs',
    # boto3 and moto required for running EC2 tests
    'boto3',
    'moto',
    "Markdown>=3.0.0",
    'parameterized',
    'lz4',
]

# NOTE: packages in the 'bundle' extra will be pinned to
#       Buildbot's exact version
# see setup.py
'bundle' = [
    "buildbot-www",
    "buildbot-worker",
    "buildbot-waterfall-view",
    "buildbot-console-view",
    "buildbot-grid-view",
]

tls = [
    'Twisted[tls]',
    # There are bugs with extras inside extras:
    # <https://github.com/pypa/pip/issues/3516>
    # so we explicitly include Twisted[tls] dependencies.
    'pyopenssl >= 16.0.0',
    'service_identity',
    'idna >= 0.6',
]

docs = [
    'docutils>=0.16.0',
    'sphinx>=3.2.0',
    'sphinx-rtd-theme>=0.5',
    'sphinxcontrib-spelling',
    'sphinxcontrib-websupport',
    'pyenchant',
    'sphinx-jinja',
    'towncrier',
]
brotli = [
    'Brotli>=1.1.0',
]
zstd = [
    'zstandard>=0.23.0',
]
configurable = [
    'evalidate >= 2.0.0',
]

[project.urls]
Homepage = "http://buildbot.net/"

[tool.aliases]
test = "trial -m buildbot"

[tool.setuptools]
include-package-data = false

[tool.setuptools.packages.find]
include = [
    "buildbot",
    "buildbot.*",
]
exclude = [
    # skip tests for wheels (save 50% of the archive)
    "buildbot.test.fuzz*",
    "buildbot.test.integration*",
    "buildbot.test.integration.interop*",
    "buildbot.test.regressions*",
    "buildbot.test.unit*",
]

[tool.setuptools.package-data]
"buildbot.reporters.templates" = ["*.txt"]
"buildbot.db.migrations" = ["alembic.ini", "README"]
"buildbot.db.migrations.versions" = ["*.py"]
"buildbot.scripts" = ["sample.cfg", "buildbot_tac.tmpl"]
"buildbot.spec" = ["*.raml"]
"buildbot.spec.types" = ["*.raml"]
"buildbot.test.unit.test_templates_dir" = ["*.html"]
"buildbot.test.unit.test_templates_dir.plugin" = ["*.*"]
"buildbot.test.integration.pki" = ["*.*"]
"buildbot.test.integration.pki.ca" = ["*.*"]

[tool.setuptools_scm]
root = ".."
version_file = "buildbot/_version.py"
local_scheme = "no-local-version"
